// ideemlatam/src/interfaces/teacher.ts
// src/interfaces/TeacherData.ts

export interface TeacherData {
    codigo: number;
    nombre: string;
    apellido: string;
    correo: string;
    nombreUsuario: string;
    telefono: string | null;
    foto: string | null;
    fechaNacimiento:string | null;
    pais: number;
    profesion: number;
    perfil: number;
    linkedin: string | null;
}